@@ -1,25 +1,20 @@
 {
   "name": "smart-chart",
-  "version": "1.0.0-beta.0",
-  "description": "A chart library base d3.js",
-  "main": "dist/index.js",
+  "version": "2.0.0-alpha.0",
+  "description": "This chart library comprises other charts, like echarts, d3 etc",
+  "main": "dist/index.umd.js",
   "module": "dist/index.es.js",
-  "unpkg": "src/index.ts",
   "types": "dist/index.d.ts",
-  "directories": {
-    "lib": "lib"
-  },
   "files": [
-    "README.md",
     "LICENSE",
-    "package.json",
-    "dist",
-    "src"
+    "dist"
   ],
   "scripts": {
-    "clean": "gulp clean",
-    "build": "npm run clean && gulp build",
-    "start": "http-server ./ -p 3000"
+    "build": "rollup -c",
+    "prebuild": "rm -rf dist/*",
+    "lint": "eslint 'src/**/*.ts'",
+    "lint:fix": "npm run lint -- --fix",
+    "prepublishOnly": "npm run build"
   },
   "repository": {
     "type": "git",
@@ -27,7 +22,8 @@
   },
   "keywords": [
     "d3.js",
-    "chart"
+    "chart",
+    "echarts"
   ],
   "author": "TaylorPzreal <taylorpzreal@gmail.com>",
   "license": "MIT",
@@ -36,30 +32,22 @@
   },
   "homepage": "https://github.com/TaylorPzreal/smart-chart#readme",
   "devDependencies": {
-    "@angular/common": "^4.4.6",
-    "@angular/compiler": "^4.4.6",
-    "@angular/compiler-cli": "^4.4.6",
-    "@angular/core": "^4.4.6",
-    "@angular/platform-browser": "^4.4.6",
-    "@types/d3": "^4.11.0",
-    "@types/node": "^8.0.46",
-    "del": "^3.0.0",
-    "gulp": "^3.9.1",
-    "gulp-copy": "^1.0.1",
-    "node-sass": "^4.5.3",
-    "rollup": "^0.50.0",
-    "rollup-plugin-angular": "^0.5.3",
-    "rollup-plugin-commonjs": "^8.2.4",
-    "rollup-plugin-node-resolve": "^3.0.0",
-    "rollup-plugin-typescript": "^0.8.1",
-    "rollup-plugin-uglify": "^2.0.1",
-    "rollup-watch": "^4.3.1",
-    "rxjs": "^5.5.0",
-    "typescript": "^2.5.3",
-    "vinyl-paths": "^2.1.0",
-    "zone.js": "^0.8.18"
+    "@babel/core": "7.11.5",
+    "@babel/preset-env": "7.11.5",
+    "@types/echarts": "4.6.5",
+    "@typescript-eslint/eslint-plugin": "4.0.1",
+    "@typescript-eslint/parser": "4.0.1",
+    "babel-eslint": "10.1.0",
+    "echarts": "4.9.0",
+    "eslint": "7.8.1",
+    "rollup": "2.26.9",
+    "rollup-plugin-babel": "4.4.0",
+    "rollup-plugin-terser": "7.0.1",
+    "rollup-plugin-typescript2": "0.27.2",
+    "typescript": "4.0.2"
   },
-  "dependencies": {
-    "d3": "^4.11.0"
-  }
+  "peerDependencies": {
+    "echats": "4^"
+  },
+  "dependencies": {}
 }
